41 - Добавил возможность удалять список и всё его содержимое. А так же возможность просто очистить список (удалить все элементы из списка), не удаляя его.

+++ app/src/main/java/com/example/shoppinglist/db/MainViewModel.kt
@@ -53,8 +53,9 @@ class MainViewModel(database: MainDataBase) : ViewModel() {
         dao.deleteNote(id)
     }
 
-    fun deleteShopListName(id: Int) = viewModelScope.launch {
-        dao.deleteShopListName(id)
+    fun deleteShopList(id: Int, deleteList: Boolean) = viewModelScope.launch {
+        if (deleteList) dao.deleteShopListName(id)
+        dao.deleteShopItemsByListId(id) //lesson 41
     }
